Preparations for Tracing Example
