+++ example_tracing.sl
@@ -11,12 +11,14 @@ function test(previous){
 function main() { 
   iteration = 0;
   value = 0;
-  setTracingEnabled(1==1);
   while (iteration < 20) {
     time = nanoTime();
     value = test(value);
     diff = (nanoTime() - time);
     iteration = iteration + 1;
     println("iteration: " + iteration + " time:" + (diff / 1000000) + "ms");
+	if (iteration == 10) {
+	  enableTracing();
+	}
   }
 }
